pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.poetry]
name = "hcaptcha-solver"
version = "0.2.14"
authors = ["xTerradon"]
description = "A lightweight python library for solving hCaptcha challenges with selenium."
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha"
]

[tool.poetry.dependencies]
python = "^3.7"
selenium = "*" 
torch = "*"
pillow = "*"
numpy = "*"
requests = "*"
unidecode = "*"
matplotlib = "*"

[tool.poetry.urls]
Homepage = "https://github.com/xTerradon/hcaptcha-solver"
